+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add latitude and longtitude to Location
 - Add expectedremuneration, jobtitle, language, skill and yearsexperience to AnnotationContentType
 - re-add DataPoint.simlarTo
+- Add `exclude` parameter to /documents query
 
 ### Changed
 - Updated endpoints for old v2 and newer v3 to point to the correct places.